api-tests/requirements.txt
click==7.1.2
behave
pyhamcrest
allure-behave
requests